Restructured webserver Added Makefile to automate deploy and tests
