+++ deploy/variables.tf
@@ -2,3 +2,13 @@ variable "do_token" {
   description = "DigitalOcean API token"
   type        = string
 }
+
+variable "go_server_user" {
+  description = "Webserver username"
+  type        = string
+}
+
+variable "go_server_pass" {
+  description = "Webserver password"
+  type        = string
+}
