+++ scripts/update_image.sh
